[AN] refactor(#299): StudyListResponse 패키지 위치 변경
